+++ AmazonChimeSDK/AmazonChimeSDK/internal/audio/protocols/AudioClientProtocol.swift
@@ -64,6 +64,22 @@ import Foundation
                       audioMode: AudioModeInternal,
                       enableAudioRedundancy: Bool) -> audio_client_status_t
 
+    // swiftlint:disable function_parameter_count variable_name
+    func startSession(_ host: String!,
+                      basePort port: Int,
+                      callId: String!,
+                      profileId: String!,
+                      microphoneMute mic_mute: Bool,
+                      speakerMute spk_mute: Bool,
+                      isPresenter presenter: Bool,
+                      sessionToken tokenString: String!,
+                      audioWsUrl: String!,
+                      callKitEnabled: Bool,
+                      appInfo: AppInfo!,
+                      audioMode: AudioModeInternal,
+                      audioDeviceCapabilities: AudioDeviceCapabilitiesInternal,
+                      enableAudioRedundancy: Bool) -> audio_client_status_t
+
     func stopSession() -> Int
 
     func isSpeakerOn() -> Bool